refactor(painters/dom): share image clip path reader
